requirements.txt
customtkinter
librosa
numpy
ruamel.yaml
matplotlib
pyworld
soundfile
pydub
requests
tkinterdnd2
pyglet
onnxruntime